Extract ObjectStoreTests into class of it's own
